clean the tikv sender and batcher
