+++ br/pkg/restore/snap_client/pipeline_items.go
@@ -16,17 +16,14 @@ package snapclient
 
 import (
 	"context"
-	"sync"
 	"time"
 
 	"github.com/pingcap/errors"
 	backuppb "github.com/pingcap/kvproto/pkg/brpb"
 	"github.com/pingcap/log"
 	"github.com/pingcap/tidb/br/pkg/glue"
-	"github.com/pingcap/tidb/br/pkg/logutil"
 	"github.com/pingcap/tidb/br/pkg/metautil"
 	restoreutils "github.com/pingcap/tidb/br/pkg/restore/utils"
-	"github.com/pingcap/tidb/br/pkg/rtree"
 	"github.com/pingcap/tidb/br/pkg/storage"
 	"github.com/pingcap/tidb/br/pkg/summary"
 	"github.com/pingcap/tidb/pkg/domain/infosync"
@@ -45,49 +42,12 @@ const defaultChannelSize = 1024
 // checksum tasks.
 const defaultChecksumConcurrency = 64
 
-// TableSink is the 'sink' of restored data by a sender.
-type TableSink interface {
-	EmitTables(tables ...CreatedTable)
-	EmitError(error)
-	Close()
-}
-
-type chanTableSink struct {
-	outCh chan<- []CreatedTable
-	errCh chan<- error
-}
-
-func (sink chanTableSink) EmitTables(tables ...CreatedTable) {
-	sink.outCh <- tables
-}
-
-func (sink chanTableSink) EmitError(err error) {
-	sink.errCh <- err
-}
-
-func (sink chanTableSink) Close() {
-	// ErrCh may has multi sender part, don't close it.
-	close(sink.outCh)
-}
-
 // CreatedTable is a table created on restore process,
 // but not yet filled with data.
 type CreatedTable struct {
 	RewriteRule *restoreutils.RewriteRules
 	Table       *model.TableInfo
 	OldTable    *metautil.Table
-}
-
-func defaultOutputTableChan() chan *CreatedTable {
-	return make(chan *CreatedTable, defaultChannelSize)
-}
-
-// TableWithRange is a CreatedTable that has been bind to some of key ranges.
-type TableWithRange struct {
-	CreatedTable
-
-	// Range has been rewrited by rewrite rules.
-	Range []rtree.Range
 }
 
 type TableIDWithFiles struct {
@@ -100,249 +60,8 @@ type TableIDWithFiles struct {
 	RewriteRules *restoreutils.RewriteRules
 }
 
-// BatchSender is the abstract of how the batcher send a batch.
-type BatchSender interface {
-	// PutSink sets the sink of this sender, user to this interface promise
-	// call this function at least once before first call to `RestoreBatch`.
-	PutSink(sink TableSink)
-	// RestoreBatch will send the restore request.
-	RestoreBatch(ranges DrainResult)
-	Close()
-}
-
-// TiKVRestorer is the minimal methods required for restoring.
-// It contains the primitive APIs extract from `restore.Client`, so some of arguments may seem redundant.
-// Maybe TODO: make a better abstraction?
-type TiKVRestorer interface {
-	// SplitRanges split regions implicated by the ranges and rewrite rules.
-	// After spliting, it also scatters the fresh regions.
-	SplitRanges(ctx context.Context,
-		ranges []rtree.Range,
-		updateCh glue.Progress,
-		isRawKv bool) error
-	// RestoreSSTFiles import the files to the TiKV.
-	RestoreSSTFiles(ctx context.Context,
-		tableIDWithFiles []TableIDWithFiles,
-		updateCh glue.Progress) error
-}
-
-type tikvSender struct {
-	client TiKVRestorer
-
-	updateCh glue.Progress
-
-	sink TableSink
-	inCh chan<- DrainResult
-
-	wg *sync.WaitGroup
-
-	tableWaiters *sync.Map
-}
-
-func (b *tikvSender) PutSink(sink TableSink) {
-	// don't worry about visibility, since we will call this before first call to
-	// RestoreBatch, which is a sync point.
-	b.sink = sink
-}
-
-func (b *tikvSender) RestoreBatch(ranges DrainResult) {
-	log.Info("restore batch: waiting ranges", zap.Int("range", len(b.inCh)))
-	b.inCh <- ranges
-}
-
-// NewTiKVSender make a sender that send restore requests to TiKV.
-func NewTiKVSender(
-	ctx context.Context,
-	cli TiKVRestorer,
-	updateCh glue.Progress,
-	splitConcurrency uint,
-) (BatchSender, error) {
-	inCh := make(chan DrainResult, defaultChannelSize)
-	midCh := make(chan drainResultAndDone, defaultChannelSize)
-
-	sender := &tikvSender{
-		client:       cli,
-		updateCh:     updateCh,
-		inCh:         inCh,
-		wg:           new(sync.WaitGroup),
-		tableWaiters: new(sync.Map),
-	}
-
-	sender.wg.Add(2)
-	go sender.splitWorker(ctx, inCh, midCh, splitConcurrency)
-	outCh := make(chan drainResultAndDone, defaultChannelSize)
-	// block on splitting and scattering regions.
-	// in coarse-grained mode, wait all regions are split and scattered is
-	// no longer a time-consuming operation, then we can batch download files
-	// as much as enough and reduce the time of blocking restore.
-	go sender.blockPipelineWorker(ctx, midCh, outCh)
-	go sender.restoreWorker(ctx, outCh)
-	return sender, nil
-}
-
-func (b *tikvSender) Close() {
-	close(b.inCh)
-	b.wg.Wait()
-	log.Debug("tikv sender closed")
-}
-
-type drainResultAndDone struct {
-	result DrainResult
-	done   func()
-}
-
-func (b *tikvSender) blockPipelineWorker(ctx context.Context,
-	inCh <-chan drainResultAndDone,
-	outCh chan<- drainResultAndDone,
-) {
-	defer close(outCh)
-	res := make([]drainResultAndDone, 0, defaultChannelSize)
-	for dr := range inCh {
-		res = append(res, dr)
-	}
-
-	for _, dr := range res {
-		select {
-		case <-ctx.Done():
-			return
-		default:
-			outCh <- dr
-		}
-	}
-}
-
-func (b *tikvSender) splitWorker(ctx context.Context,
-	ranges <-chan DrainResult,
-	next chan<- drainResultAndDone,
-	concurrency uint,
-) {
-	defer log.Debug("split worker closed")
-	eg, ectx := errgroup.WithContext(ctx)
-	defer func() {
-		b.wg.Done()
-		if err := eg.Wait(); err != nil {
-			b.sink.EmitError(err)
-		}
-		close(next)
-		log.Info("TiKV Sender: split worker exits.")
-	}()
-
-	start := time.Now()
-	defer func() {
-		elapsed := time.Since(start)
-		summary.CollectDuration("split region", elapsed)
-	}()
-
-	pool := tidbutil.NewWorkerPool(concurrency, "split")
-	for {
-		select {
-		case <-ectx.Done():
-			return
-		case result, ok := <-ranges:
-			if !ok {
-				return
-			}
-			// When the batcher has sent all ranges from a table, it would
-			// mark this table 'all done'(BlankTablesAfterSend), and then we can send it to checksum.
-			//
-			// When there a sole worker sequentially running those batch tasks, everything is fine, however,
-			// in the context of multi-workers, that become buggy, for example:
-			// |------table 1, ranges 1------|------table 1, ranges 2------|
-			// The batcher send batches: [
-			//		{Ranges: ranges 1},
-			// 		{Ranges: ranges 2, BlankTablesAfterSend: table 1}
-			// ]
-			// And there are two workers runs concurrently:
-			// 		worker 1: {Ranges: ranges 1}
-			//      worker 2: {Ranges: ranges 2, BlankTablesAfterSend: table 1}
-			// And worker 2 finished its job before worker 1 done. Note the table wasn't restored fully,
-			// hence the checksum would fail.
-			done := b.registerTableIsRestoring(result.TablesToSend)
-			pool.ApplyOnErrorGroup(eg, func() error {
-				err := b.client.SplitRanges(ectx, result.Ranges, b.updateCh, false)
-				if err != nil {
-					log.Error("failed on split range", rtree.ZapRanges(result.Ranges), zap.Error(err))
-					return err
-				}
-				next <- drainResultAndDone{
-					result: result,
-					done:   done,
-				}
-				return nil
-			})
-		}
-	}
-}
-
-// registerTableIsRestoring marks some tables as 'current restoring'.
-// Returning a function that mark the restore has been done.
-func (b *tikvSender) registerTableIsRestoring(ts []CreatedTable) func() {
-	wgs := make([]*sync.WaitGroup, 0, len(ts))
-	for _, t := range ts {
-		i, _ := b.tableWaiters.LoadOrStore(t.Table.ID, new(sync.WaitGroup))
-		wg := i.(*sync.WaitGroup)
-		wg.Add(1)
-		wgs = append(wgs, wg)
-	}
-	return func() {
-		for _, wg := range wgs {
-			wg.Done()
-		}
-	}
-}
-
-// waitTablesDone block the current goroutine,
-// till all tables provided are no more ‘current restoring’.
-func (b *tikvSender) waitTablesDone(ts []CreatedTable) {
-	for _, t := range ts {
-		wg, ok := b.tableWaiters.LoadAndDelete(t.Table.ID)
-		if !ok {
-			log.Panic("bug! table done before register!",
-				zap.Any("wait-table-map", b.tableWaiters),
-				zap.Stringer("table", t.Table.Name))
-		}
-		wg.(*sync.WaitGroup).Wait()
-	}
-}
-
-func (b *tikvSender) restoreWorker(ctx context.Context, ranges <-chan drainResultAndDone) {
-	eg, ectx := errgroup.WithContext(ctx)
-	defer func() {
-		log.Info("TiKV Sender: restore worker prepare to close.")
-		if err := eg.Wait(); err != nil {
-			b.sink.EmitError(err)
-		}
-		b.sink.Close()
-		b.wg.Done()
-		log.Info("TiKV Sender: restore worker exits.")
-	}()
-	for {
-		select {
-		case <-ectx.Done():
-			return
-		case r, ok := <-ranges:
-			if !ok {
-				return
-			}
-
-			files := r.result.Files()
-			// There has been a worker in the `RestoreSSTFiles` procedure.
-			// Spawning a raw goroutine won't make too many requests to TiKV.
-			eg.Go(func() error {
-				e := b.client.RestoreSSTFiles(ectx, files, b.updateCh)
-				if e != nil {
-					log.Error("restore batch meet error", logutil.ShortError(e), zapTableIDWithFiles(files))
-					r.done()
-					return e
-				}
-				log.Info("restore batch done", rtree.ZapRanges(r.result.Ranges), zapTableIDWithFiles(files))
-				r.done()
-				b.waitTablesDone(r.result.BlankTablesAfterSend)
-				b.sink.EmitTables(r.result.BlankTablesAfterSend...)
-				return nil
-			})
-		}
-	}
+func defaultOutputTableChan() chan *CreatedTable {
+	return make(chan *CreatedTable, defaultChannelSize)
 }
 
 func concurrentHandleTablesCh(
@@ -384,83 +103,6 @@ func concurrentHandleTablesCh(
 			})
 		}
 	}
-}
-
-// GoValidateFileRanges validate files by a stream of tables and yields
-// tables with range.
-func (rc *SnapClient) GoValidateFileRanges(
-	ctx context.Context,
-	tableStream <-chan CreatedTable,
-	fileOfTable map[int64][]*backuppb.File,
-	splitSizeBytes, splitKeyCount uint64,
-	errCh chan<- error,
-) <-chan TableWithRange {
-	// Could we have a smaller outCh size?
-	outCh := make(chan TableWithRange, len(fileOfTable))
-	go func() {
-		defer close(outCh)
-		defer log.Info("all range generated")
-		for {
-			select {
-			case <-ctx.Done():
-				errCh <- ctx.Err()
-				return
-			case t, ok := <-tableStream:
-				if !ok {
-					return
-				}
-				files := fileOfTable[t.OldTable.Info.ID]
-				if partitions := t.OldTable.Info.Partition; partitions != nil {
-					log.Debug("table partition",
-						zap.Stringer("database", t.OldTable.DB.Name),
-						zap.Stringer("table", t.Table.Name),
-						zap.Any("partition info", partitions),
-					)
-					for _, partition := range partitions.Definitions {
-						files = append(files, fileOfTable[partition.ID]...)
-					}
-				}
-				for _, file := range files {
-					err := restoreutils.ValidateFileRewriteRule(file, t.RewriteRule)
-					if err != nil {
-						errCh <- err
-						return
-					}
-				}
-				// Merge small ranges to reduce split and scatter regions.
-				ranges, stat, err := restoreutils.MergeAndRewriteFileRanges(
-					files, t.RewriteRule, splitSizeBytes, splitKeyCount)
-				if err != nil {
-					errCh <- err
-					return
-				}
-				log.Info("merge and validate file",
-					zap.Stringer("database", t.OldTable.DB.Name),
-					zap.Stringer("table", t.Table.Name),
-					zap.Int("Files(total)", stat.TotalFiles),
-					zap.Int("File(write)", stat.TotalWriteCFFile),
-					zap.Int("File(default)", stat.TotalDefaultCFFile),
-					zap.Int("Region(total)", stat.TotalRegions),
-					zap.Int("Regoin(keys avg)", stat.RegionKeysAvg),
-					zap.Int("Region(bytes avg)", stat.RegionBytesAvg),
-					zap.Int("Merged(regions)", stat.MergedRegions),
-					zap.Int("Merged(keys avg)", stat.MergedRegionKeysAvg),
-					zap.Int("Merged(bytes avg)", stat.MergedRegionBytesAvg))
-
-				tableWithRange := TableWithRange{
-					CreatedTable: t,
-					Range:        ranges,
-				}
-				log.Debug("sending range info",
-					zap.Stringer("table", t.Table.Name),
-					zap.Int("files", len(files)),
-					zap.Int("range size", len(ranges)),
-					zap.Int("output channel size", len(outCh)))
-				outCh <- tableWithRange
-			}
-		}
-	}()
-	return outCh
 }
 
 // GoValidateChecksum forks a goroutine to validate checksum after restore.
